Trigger correction
The trigger is now based on a moving average (acceleration values are buffered in a table and the average of the table is compared to the desired threshold). Every acceleration value is a square root of a sum of SQUARES.
The vector VECTOR_LINEARACCEL is used instead of VECTOR_ACCELEROMETER as it already accounts for gravity.