Redundant trigger, altimeter calib, accelerometer range
The trigger can now be activated by an overpressure within the payload bay - it should be checked that the closure of the nosecone doesn't trigger it.

The altimeter is calibrated with the pressure measured 1 minute before liftoff.

The libraries of the accelerometer have been modified so the G range can easily be modified during the setup.