Almost ready to merge !
The frequency is good. The way of saving the files is robust  and they contain all the necessary information for a given pin at a given moment (A* Timeinµs Amp)

Some values are still discarded because of the time taken to copy the buffer. This can be solved by using 2 buffers again but with an interrupt to equilibrate their content.